mock requests.get response for socrata tests
