+++ dcpy/test/conftest.py
@@ -1,5 +1,6 @@
 import csv
 from datetime import datetime
+import json
 from moto import mock_aws
 import os
 from pathlib import Path
@@ -15,6 +16,7 @@ from dcpy.lifecycle.builds import plan
 
 TEST_BUCKET = "test-bucket"
 TEST_BUCKETS = [TEST_BUCKET, publishing.BUCKET, packaging.BUCKET, recipes.BUCKET]
+RESOURCES = Path(__file__).parent / "resources"
 
 
 @pytest.fixture(scope="function")
@@ -130,3 +132,42 @@ def create_temp_filesystem(mock_data_constants):
     except Exception as e:
         f"❌ Unable to remove {data_path} after running tests"
         raise e
+
+
+def mock_request_get(url):
+    """
+    Mocks calls to request.get
+
+    To use, annotate test with
+    @mock.patch("requests.get", side_effect=mock_request_get)
+    If you do this, the first argument of the test MUST be a sort of "dummy" variable. This is
+    meant more for if the mocked variable is a Class than a function. But if you attempt to use
+    a fixture as the first argument to a function with this annotation, it will not actually be accessed.
+
+    To use, add a new path and filename to the `test_files` dictionary and add the appropriate file
+    (containing a plain string response) to resources/mocked_responses
+    """
+
+    class MockResponse:
+        def __init__(self, content: bytes):
+            self.content = content
+
+        def json(self):
+            return json.loads(self.content)
+
+        def raise_for_status(self):
+            pass
+
+    test_files = {
+        "https://www.bklynlibrary.org/locations/json": "bpl_libraries.json",
+        "https://data.cityofnewyork.us/api/views/w7w3-xahh/rows.csv": "dca_operatingbusinesses.csv",
+        "https://s-media.nyc.gov/agencies/dcp/assets/files/zip/data-tools/bytes/pad_24a.zip": "pad_24a.zip",
+        "https://health.data.ny.gov/api/views/izta-vnpq/rows.csv": "nysdoh_nursinghomes.csv",
+        "https://data.cityofnewyork.us/api/views/w7w3-xahh.json": "dca_operatingbusinesses_metadata.json",
+    }
+
+    if url not in test_files:
+        raise Exception(f"Url {url} has not been configured with test data")
+
+    with open(RESOURCES / "mocked_responses" / test_files[url], "rb") as file:
+        return MockResponse(file.read())
