--- dcpy/test/connectors/socrata/resources/sample_api_response.json
@@ -1,116 +1,0 @@
-{
-    "id" : "q663-gvx6",
-    "name" : "Outcome of Preventive Cases Closed By Borough And CD (CY 2014)",
-    "assetType" : "file",
-    "attribution" : "Administration for Children's Services (ACS)",
-    "averageRating" : 0,
-    "blobFilename" : "Outcome of Preventive Cases Closed CY2014 by Borough and CD.xlsx",
-    "blobFileSize" : 105695,
-    "blobId" : "42624a76-d9ec-41db-a39a-1617d0ea1b41",
-    "blobMimeType" : "application/vnd.ms-excel; charset=binary",
-    "category" : "Social Services",
-    "createdAt" : 1465231616,
-    "description" : "this report was developed in response to a request posted on Open Data site. There is no plan for updates. This report provides the number and percent of child welfare preventive cases closed that had an indicated investigation or foster care placement within six months. by comunnity district and brourough of a family.",
-    "displayType" : "blob",
-    "downloadCount" : 661,
-    "hideFromCatalog" : false,
-    "hideFromDataJson" : false,
-    "indexUpdatedAt" : 1536600759,
-    "newBackend" : false,
-    "numberOfComments" : 0,
-    "oid" : 21465349,
-    "provenance" : "official",
-    "publicationAppendEnabled" : false,
-    "publicationDate" : 1465231616,
-    "publicationGroup" : 11169728,
-    "publicationStage" : "published",
-    "rowsUpdatedAt" : 1465231616,
-    "tableId" : 11169728,
-    "totalTimesRated" : 0,
-    "viewCount" : 554,
-    "viewLastModified" : 1536605680,
-    "viewType" : "blobby",
-    "approvals" : [ {
-      "reviewedAt" : 1465232023,
-      "reviewedAutomatically" : true,
-      "state" : "approved",
-      "submissionId" : 1066517,
-      "submissionObject" : "public_audience_request",
-      "submissionOutcome" : "change_audience",
-      "submittedAt" : 1465232023,
-      "targetAudience" : "public",
-      "workflowId" : 2285,
-      "submissionDetails" : {
-        "permissionType" : "READ"
-      },
-      "submissionOutcomeApplication" : {
-        "failureCount" : 0,
-        "status" : "success"
-      },
-      "submitter" : {
-        "id" : "5fuc-pqz2",
-        "displayName" : "NYC OpenData"
-      }
-    } ],
-    "clientContext" : {
-      "clientContextVariables" : [ ],
-      "inheritedVariables" : { }
-    },
-    "columns" : [ ],
-    "grants" : [ {
-      "inherited" : false,
-      "type" : "viewer",
-      "flags" : [ "public" ]
-    } ],
-    "metadata" : {
-      "attachments" : [ {
-        "filename" : "Data_Dictionary_Outcome_of_preventive_cases_closed_by_CD.xlsx",
-        "assetId" : "7d878551-e60b-40ac-9bce-f6068f2deee0",
-        "name" : "Data_Dictionary_Outcome_of_preventive_cases_closed_by_CD.xlsx"
-      } ],
-      "custom_fields" : {
-        "Update" : {
-          "Automation" : "No",
-          "Date Made Public" : "6/6/2016",
-          "Update Frequency" : "As needed"
-        },
-        "Legislative Compliance" : {
-          "Geo-coding Attached?" : ""
-        },
-        "Dataset Information" : {
-          "Agency" : "Administration for Children's Services (ACS)"
-        }
-      },
-      "availableDisplayTypes" : [ "blob" ],
-      "rowLabel" : "Row",
-      "renderTypeConfig" : {
-        "visible" : {
-          "blob" : true
-        }
-      }
-    },
-    "owner" : {
-      "id" : "5fuc-pqz2",
-      "displayName" : "NYC OpenData",
-      "profileImageUrlLarge" : "/api/users/5fuc-pqz2/profile_images/LARGE",
-      "profileImageUrlMedium" : "/api/users/5fuc-pqz2/profile_images/THUMB",
-      "profileImageUrlSmall" : "/api/users/5fuc-pqz2/profile_images/TINY",
-      "screenName" : "NYC OpenData",
-      "type" : "interactive",
-      "flags" : [ "acceptedEula", "mayBeStoriesCoOwner" ]
-    },
-    "query" : { },
-    "rights" : [ "read" ],
-    "tableAuthor" : {
-      "id" : "5fuc-pqz2",
-      "displayName" : "NYC OpenData",
-      "profileImageUrlLarge" : "/api/users/5fuc-pqz2/profile_images/LARGE",
-      "profileImageUrlMedium" : "/api/users/5fuc-pqz2/profile_images/THUMB",
-      "profileImageUrlSmall" : "/api/users/5fuc-pqz2/profile_images/TINY",
-      "screenName" : "NYC OpenData",
-      "type" : "interactive",
-      "flags" : [ "acceptedEula", "mayBeStoriesCoOwner" ]
-    },
-    "tags" : [ "preventive services", "foster care", "child abuse" ],
-    "flags" : [ "default", "ownerMayBeContacted" ]
-  }
