+++ dcpy/test/resources/mocked_responses/dca_operatingbusinesses_metadata.json
