+++ dcpy/test/resources/mocked_responses/pad_24a.zip
